Fix null deref when glGetString(GL_EXTENSIONS) == NULL
When using a core profile OpenGL 4.1 context on macOS,
EXT_texture_sRGB is not listed. This causes Horde3D to
fall back to OpenGL 2.1, but without switching the context profile.

I'm not clear on the semantics of an OpenGL 2.1 core profile
context on macOS or how that's possible, but
glGetString(GL_EXTENSIONS) returns NULL.