Support meshes with line list primitive type
This patch is based on code from attilaz posted on the forum long time ago.
Triangle strip might not seem to be that useful, but I need support for drawing lines from regular API
(for technical drawing). This allows to use the regular API to draw lines.

It changes slightly the tesselation API as it uses now the primitive type to specify that we want
Patches instead of a seperate flag.