Support meshes with line list primitive type
This patch is based on code from attilaz posted on on the forum long time ago.
Triangle strip might seem to be that useful, but I need support for drawing lines from regular API
(for technical drawing).
It uses the primitive type member to issue patches for tesselation instead of separate flag.