Allow to create render buffers with mipmaps
When a render buffer is applied as a texture to a 3D model, there can be heavy aliasing artefacts
if no mipmapping is applied (for example when rendering test to be applied on a model).
This patch allowd to automatically update mipmap levels at the end of drawing to the render buffer.