Replace `atof` by locale independent `toFloat`
When the application's locale is changed to one that uses comma as decimal separator, then
the xml parsing doesn't work correctly anymore (fractional part after comma is ignored).
German and French locales are using comma separator. Typically the default locale is the
"C" locale, but for example a Qt application under linux gets the system locale.
The application's locale is not always controlled and it's better to parse xml with functions
that are independent of the global `setlocale`.