Support texture with defined mip map level (eg. not necessarily full mip map down to 1x1)
Previously, loading a dds or ktx file that has partial mip map definition would result
in incomplete texture that appears black in shader.

This patch essentially maintains the maximum mipmap level for a texture and sets
GL_TEXTURE_MAX_LEVEL when creating the texture.
The GL_TEXTURE_MAX_LEVEL parameter is part of the texture parameters (not sampler state),
therefore it should be sufficient to set it when texture is created.
(see: https://www.khronos.org/opengl/wiki/Texture)

(Note: having partial mip map is not completely exotic; for instance this is commonly\
used in irradiance map for PBR rendering)